Made Quick Menu options using keyboard
Added Space to start game and escape to go to menu (will be pause when i
create the pause menu)

Also tested to see if the keyboard could switch the Game.State (which
the mouse isnt)

I think the mouse error is that it either isnt being called correctly or
that it isnt registering the correct area that is being clicked.  Either
way when i have time later today im going to completely redo the mouse
stuff.